a lot of improvment of code
now I have to go ahead from main_menu() -> case 'y'
