+++ .vscode/settings.json
@@ -1,0 +1,10 @@
+{
+    "files.associations": {
+        "unistd.h": "c",
+        "stdlib.h": "c",
+        "stdarg.h": "c",
+        "locale.h": "c",
+        "environments.h": "c",
+        "unistd_ext.h": "c"
+    }
+}
