+++ rover.1
@@ -1,4 +1,4 @@
-.TH ROVER 1 2020-06-04 rover\-1.0.1
+.TH ROVER 2 2022-12-31 rover\-2.0.1
 .SH NAME
 rover \- file browser for the terminal
 .SH SYNOPSIS
@@ -75,14 +75,14 @@ pointed to the "source" directory of the operation and then issue the command on
 another tab that is pointed to the "destination" directory.
 .SH COMMANDS
 .TP
-.B q
+.B ESC
 Quit rover.
 .TP
-.B j/k
-Move cursor down/up.
+.B arrow up/down
+Move cursor up/down.
 .TP
-.B J/K
-Move cursor down/up 10 lines.
+.B Page up/down
+Move cursor up/down 10 lines.
 .TP
 .B g/G
 Move cursor to top/bottom of listing.
