+++ src/config.h
@@ -7,7 +7,6 @@
 #define _XOPEN_SOURCE_EXTENDED
 #define _FILE_OFFSET_BITS 64
 
-#include <stdlib.h>
 #include <stdint.h>
 #include <ctype.h>
 #include <wchar.h>
@@ -29,44 +28,7 @@
 #include <curses.h>
 #include <stdbool.h>
 
-#define RV_VERSION "1.0.1"
-
-/* CTRL+X: "^X"
-   ALT+X: "M-X" */
-#define RVK_QUIT        "q"
-#define RVK_HELP        "?"
-#define RVK_DOWN        "j"
-#define RVK_UP          "k"
-#define RVK_JUMP_DOWN   "J"
-#define RVK_JUMP_UP     "K"
-#define RVK_JUMP_TOP    "g"
-#define RVK_JUMP_BOTTOM "G"
-#define RVK_CD_DOWN     "l"
-#define RVK_CD_UP       "h"
-#define RVK_HOME        "H"
-#define RVK_TARGET      "t"
-#define RVK_COPY_PATH   "y"
-#define RVK_PASTE_PATH  "p"
-#define RVK_REFRESH     "r"
-#define RVK_SHELL       "^M"
-#define RVK_VIEW        " "
-#define RVK_EDIT        "e"
-#define RVK_OPEN        "o"
-#define RVK_SEARCH      "/"
-#define RVK_TG_FILES    "f"
-#define RVK_TG_DIRS     "d"
-#define RVK_TG_HIDDEN   "s"
-#define RVK_NEW_FILE    "n"
-#define RVK_NEW_DIR     "N"
-#define RVK_RENAME      "R"
-#define RVK_TG_EXEC     "E"
-#define RVK_DELETE      "D"
-#define RVK_TG_MARK     "m"
-#define RVK_INVMARK     "M"
-#define RVK_MARKALL     "a"
-#define RVK_MARK_DELETE "X"
-#define RVK_MARK_COPY   "C"
-#define RVK_MARK_MOVE   "V"
+#define RV_VERSION "2.0.1"
 
 /* Special symbols used by the TUI. See <curses.h> for available constants. */
 #define RVS_SCROLLBAR ACS_CKBOARD
@@ -79,26 +41,12 @@
 #define RVP_NEW_DIR  RV_PROMPT("new dir")
 #define RVP_RENAME   RV_PROMPT("rename")
 
-/* Number of entries to jump on RVK_JUMP_DOWN and RVK_JUMP_UP. */
-#define RV_JUMP 10
-
 /* Default listing view flags.
    May include SHOW_FILES, SHOW_DIRS and SHOW_HIDDEN. */
 #define RV_FLAGS SHOW_FILES | SHOW_DIRS
 
 /* Optional macro to be executed when a batch operation finishes. */
 #define RV_ALERT() beep()
-
-/* Shell used to launch external  programs.
-   Defining   this  macro   will  force   Rover  to   launch  external
-   programs with  `sh -c  "$EXTERNAL_PROGRAM [arg]"`. This  gives more
-   flexibility,  allowing command-line  arguments  to  be embedded  in
-   environment variables  (e.g. PAGER="less  -N"). On the  other hand,
-   this requires the presence of a  shell and will spawn an additional
-   process each time an external  program is invoked. Leave this macro
-   undefined if you prefer external  programs to be launched with just
-   `$EXTERNAL_PROGRAM [arg]`. */
-#define RV_SHELL "/bin/sh"
 
 /*  This signal is not defined by POSIX, but should be
    present on all systems that have resizable terminals. */
@@ -229,17 +177,11 @@ void mark_none(Marks *marks);
 void add_mark(Marks *marks, char *dirpath, char *entry);
 void del_mark(Marks *marks, char *entry);
 void free_marks(Marks *marks);
-void handle_usr1(int sig);
-void handle_winch(int sig);
-void enable_handlers();
-void disable_handlers();
 void reload();
 void sync_signals();
 int rover_getch();
 int rover_get_wch(wint_t *wch);
-void spawn(char **args);
 void shell_escaped_cat(char *buf, char *str, size_t n);
-int open_with_env(char *program, char *path);
 void update_view();
 int rowcmp(const void *a, const void *b);
 int ls(Row **rowsp, uint8_t flags);
